CI: commit build results
The CI script now creates a new branch (`ci-of-$HASH`) for each build,
whose parent is the source commit and which adds the log output of
building it. The commit message says whether it passed or failed.

You can `git fetch ci` to pull in all the logs and view them locally.
`gitk -all` will show the build results as branches from the source
commits.

If the results branch already exists it will not rebuild.