Replace Irmin.View with I9p_rw
I9p_rw simply holds a mutable reference to an I9p_tree.Dir (the root)
and provides convenient functions to update it. I9p_tree has been
extended to support both in-memory and on-disk files and directories.

Some advantages:

- Unlike Irmin.View, we do not keep a log of actions performed on the
  transaction. Keeping this was causing us to run out of memory.

- We can get a snapshot of a view cheaply, by simply taking a reference
  to the current root node (which is immutable). This may fix #16.

The stream watchers had to change slightly because `hash` is now async
(might need to flush to disk), although in fact this can't happen
because we don't allow watching views.

Signed-off-by: Thomas Leonard <thomas.leonard@docker.com>