Allow connections on named pipes
If built with NAMED_PIPE=enable then the binary will be able to listen on
Windows named pipes for incoming connections.

The accept/handle loops for Lwt_unix and Named_pipe.Server look a little different
so there are now 2 wrappers:

- unix_accept_forever
- named_pipe_accept_forever

Note that previously there was no Lwt_unix.close of the accepted fds, so it looks
like the server leaked connections. This has been fixed.

Signed-off-by: David Scott <dave.scott@docker.com>