Close the flow on disconnection
Previously the flow was closed prematurely because the callback
returns once the server connection to the client has been negotiated.

Recent versions of protocol-9p have an `after_disconnect` thread which
allows resource cleanup to be triggered when the 9P connection is
disconnected. This patch uses this to call Flow.close after the 9P
disconnection.

Note that the normal Lwt_unix.close should be good enough for named pipes
because they are represented as HANDLES underneath. There should be no
need to explicitly flush (as the 9P protocol has already consumed all
it wants) and no need to explicitly disconnect (as we don't intend to
reconnect the handle to any other client).

Signed-off-by: David Scott <dave.scott@docker.com>