More robust GH bridge
The current architecture is the following (where `-->` means `commit`
and `-m->` mens `merge`):

    GH <---> priv branch -m-> pub branch <--- user

The server keeps in memory a snapshot (prs and status) for both the
private and public branches, of the last time it did a synchronisation.
This helps when to compare in the both branches since last sync, and
do "the right thing" to converge.

Signed-off-by: Thomas Gazagnaire <thomas@gazagnaire.org>