Support inheriting a listening socket
When starting a set of services together it is useful to be able to
pre-bind the listening sockets, to avoid races where a `connect` fails
because one of the components hasn't yet called `listen`.

This patch adds support for arguments of the form

  --arg fd:4

where '4' would be the listening socket file descriptor created by
the parent process.

Signed-off-by: David Scott <dave.scott@docker.com>