github: do not fail when removing non-existing directories
Also do not try to remove multiple times the commit directory
corresponding to a set of build status.

This was causing the merge between the github-metadata and
github-metadata-x branches to fail and abort the current
reconciliation loop. A new (unrelated) change were needed
to kick off the reconciliation loop again.

Signed-off-by: Thomas Gazagnaire <thomas@gazagnaire.org>