github: refactor the GitHub bridge
Previously, the bridge was using two DataKit branches:

- a "private" branch which will mirror the GitHub state, populated on
  startup by GitHub API calls, and updated regularly by a separate
  webhook listener directly connected to DataKit.

- a "public" branch, regularly merged with the private one, where users
  could see the current state of GitHub and also change some parts of
  the state (for instance, update the build status of a given commit)
  to trigger appropriate GitHub API calls.

Using that scheme: conflicts could happen, (as webhooks could arrive
at the same time as a user request to change some state), the merge
strategy is handled by the DataKit server (and hence, it doesn't
understand the invariants that should be kept in the GH
state). Moreover, as the bridge doesn't have any handle on both the
incoming webhooks and the user requests, it is pretty hard for it to
keep a consistent view of the world without loosing events (either
from GH or from the users).  Finally, any conflict/error could lead to
surprising results ("oh, the GH bridge reverted the title of that PR
to something old!")  and was a sometimes unreliable.

This (rather large) patches refactor the existing code to simplify
that process.

- There is now only one branch in DataKit, where users can see the
  state of GitHub and request changes (the equivalent of the previous
  "public" branch).

- The "private" branch now only exists in the bridge memory, and is
  rebuilt from scratch on every restart. Webhooks events are kept in
  a queue and read/applied when needed (instead of arriving in
  asynchronous order at any point in time). Conflict are handled
  by letting the user specifying extended capabilities/ownership
  on every GH resources. This let for instance users the ability
  to only modify the build status starting by a given prefix, and
  nothing else.

Signed-off-by: Thomas Gazagnaire <thomas@gazagnaire.org>