ci: improve UI for viewing logs
Instead of showing all the logs at once, use a split-pane view with the
single selected log shown in the bottom pane.

This avoids transferring huge amounts of log data that the user may not
be interested in and makes navigation easier.

By default, the log shown initially is the first of:
- the first failed log
- the first pending log
- the first successful log

Rebuild buttons are now attached to logs, not jobs. This makes for a better
UI, since the user will typically be looking at a log showing a failure
and want to rebuild that step. The rebuild button is highlighted in red
for logs of failed builds.

The summary area is now more compact to avoid using up too much of the
available screen area.

Signed-off-by: Thomas Leonard <thomas.leonard@docker.com>