ci: make job target explicit
Before, Term.target returned the target for which the term was being
evaluated. However, we might want to evaluate terms in contexts without
a target too.

Now, the test configuration for a project is a function from a target
to the set of tests. This also means that the tests to perform can
depend on the target ID. For example more tests can be run on release
tags, and different tests can be run on a gh-pages branch.

Signed-off-by: Thomas Leonard <thomas.leonard@docker.com>