github: Add a `--resync-interval` option to resync the database regularly
The way it is implemented is pretty dumb at the moment, and it will not
scale out very well when we will have to handle 100s of repos (as it will
consume the API credits pretty quickly). But it should address the short-term
issues of flacky event delivery.

By default resync every hour. Use`--no-resync` to never resync.

Signed-off-by: Thomas Gazagnaire <thomas@gazagnaire.org>