ci: fix fetching of tag objects
Before, we tried to fetch each Git target to a local Git branch. That
works for PRs, branches and plain tags, but not for tag objects. Those
can only be stored in local tags.

Now, we fetch everything to a local tag first, and then copy just the
commit to the branch.

Also, added debug-level logging for child process output. This is useful
to see why the unit-tests are failing.

Signed-off-by: Thomas Leonard <thomas.leonard@docker.com>