Web-based configuration for admin user
Before, we handled initial configuration by prompting the user to choose
a password on stdin. However, this doesn't work well when using
docker-compose or docker-cloud, since stdin isn't connected.

Instead, we now print a configuration URL to the logs in this case,
which includes a one-shot token allowing the initial configuration.

Signed-off-by: Thomas Leonard <thomas.leonard@docker.com>