ci: improve form validation
Instead of showing an error page that is blank except for the error
message when an invalid form is submitted, show the form again with the
entered values still filled in and invalid fields tagged with the
errors.

Also, moved the core form logic into a new CI_form module.

Signed-off-by: Thomas Leonard <thomas.leonard@docker.com>