ci: add support for getting X.509 certificates via certbot
This is just a quick hack. It adds a script that you can run to update
the certificates. It works because certbot actually only needs to run on
port 80 and so doesn't need to disturb the CI.

I've used this to move the server's official name from datakit.ci:8446
to datakit.datakit.ci.

This PR also makes some improvements to the form system because I
originally had a web UI for this, but they seemed worth keeping anyway.

There are many ways it could be improved:

- Don't require the server to be restarted to use the new certificate.
- Run the script automatically when the certificate is due to expire.
- Report a metric saying how long the current certificate is valid for.

Signed-off-by: Thomas Leonard <thomas.leonard@docker.com>