ci: disable secure cookies when testing without https
If --web-ui is used with a plain "http" scheme, don't use secure
cookies. This may be useful for testing locally.

Note that configuring the listen address as HTTP does not affect this,
since we might have an https proxy in front of us and we still want
secure cookies in that case.

Signed-off-by: Thomas Leonard <thomas.leonard@docker.com>