Add local git bridge
Normally, we use datakit-github to monitor the state of a remote
repository on GitHub and use that as the input to the CI. When getting
started with DataKitCI it is more convenient to be able to monitor a
local Git repository.

This also extends SelfCI with a docker-compose file that uses this local
bridge to run the tests against the local repository.

Signed-off-by: Thomas Leonard <thomas.leonard@docker.com>