self-ci: test DataKit and Prometheus too
This removes the last uses of Travis.

Added `?entrypoint` to `Docker.command` to support this.

Also, switched to opam 2 for builds. This should make the tests much
faster because it doesn't try to install (and test) all the test
dependencies, recursively.

Note: opam 2 defaults to "no" in many places where opam 1 defaults to
yes. If it aborts because it's not running interactively, it returns
success. I've set `ENV OPAMYES=1` everywhere to try to avoid this.

Signed-off-by: Thomas Leonard <thomas.leonard@docker.com>