ci: add a lock to update statuses in series
Because multiple targets can be pointing to the same commit, we may try
to update a single commit's status for several targets in parallel,
leading to merge conflicts. This is a limitation of GitHub's design, so
just let the most recent update to win.

Without this, you will often see merge conflict warnings in the logs.

Signed-off-by: Thomas Leonard <thomas.leonard@docker.com>