ci: recalculate all jobs on rebuild
Before, triggering a rebuild only recalculated terms with that value in
their log history. However:

1. Terms can hide logs.
2. There is little performance benefit since as soon as the new pending
   status is written, we recalculate everything anyway.

This means that we no longer recalculate things out of order, which will
allow batching up updates. That in turn avoids the problem that the
history of a target may have results from different versions of the
input, which is confusing.

Signed-off-by: Thomas Leonard <thomas.leonard@docker.com>