github: don't ignore errors from DataKit
Lots of these functions used to ignore errors silently. This can lead
to bugs in the users of the library (e.g. we lost our 9p connection but
we think instead that the file we wanted doesn't exist). For now, I've
converted it to log errors in these cases but continue with the old
behaviour. Assuming we don't see these errors being logged, we can
change the code to raise exceptions instead.

Also, updated the client to report "Can't walk from a file" as `Not_dir.

Signed-off-by: Thomas Leonard <thomas.leonard@docker.com>