ci: make required GitHub scopes configurable
Before, we always asked for access to all public and private repos, in
case we needed to check if the user had access to a private repo. Now,
a CI can request only access to public repos if that's all it needs.

Workaround until https://github.com/mirage/ocaml-github/issues/185
is implemented.

Signed-off-by: Thomas Leonard <thomas.leonard@docker.com>